+++ src/Core/CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - Support for ECS Credentials Provider
+- Support for Cognito Identity Provider client in `AwsClientFactory`
+- Support for Cloud Watch Log client in `AwsClientFactory`
 
 ### Fixed